+++ R/btw-package.R
@@ -7,6 +7,7 @@
 #' @importFrom brio read_lines
 #' @importFrom brio write_file
 #' @importFrom brio write_lines
+#' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL